+++ apps/hardware_benchmarks/tests/db_3_3/db_3_3_generator.cpp
@@ -34,7 +34,7 @@ public:
         conv(x, y)  += u16(kernel(r.x, r.y)) * hw_input(x + r.x, y + r.y);
 
         Func hw_output("hw_output");
-        hw_output(x, y) = u8(conv(x, y));
+        hw_output(x, y) = conv(x, y);
         output(x, y) = u8(hw_output(x,y));
 
         /* THE SCHEDULE */
@@ -43,12 +43,12 @@ public:
 
           hw_output.compute_root();
 
-        conv.bound(x, 0, imgsize);
-        conv.bound(y, 0, imgsize);
-        output.bound(x, 0, imgsize);
-        output.bound(y, 0, imgsize);
-        hw_output.bound(x, 0, imgsize);
-        hw_output.bound(y, 0, imgsize);
+          conv.bound(x, 0, imgsize);
+          conv.bound(y, 0, imgsize);
+          output.bound(x, 0, imgsize);
+          output.bound(y, 0, imgsize);
+          hw_output.bound(x, 0, imgsize);
+          hw_output.bound(y, 0, imgsize);
           
           hw_output.tile(x,y, xo,yo, xi,yi, imgsize, imgsize)
             .hw_accelerate(xi, xo);
